statsd/datadog: refactor tests
