+++ stats/statsd/collector_test.go
@@ -1,0 +1,15 @@
+package statsd
+
+import (
+	"testing"
+
+	"github.com/loadimpact/k6/stats"
+	"github.com/loadimpact/k6/stats/statsd/common/testutil"
+)
+
+func TestCollector(t *testing.T) {
+	testutil.BaseTest(t, New,
+		func(_ []stats.SampleContainer, output string) string {
+			return output
+		})
+}
